handling Retry-After (with relative time only)

+++ python/download_base.py
@@ -1,16 +1,24 @@
+import re
 import select
 import sys
+import time
 from common import get_option
-from cursor_wrapper import CursorWrapper
+from host_check import HostCheck
 
-class DownloadBase(CursorWrapper):
+class DownloadBase(HostCheck):
     def __init__(self, conn, cur, single_action):
-        CursorWrapper.__init__(self, cur)
+        HostCheck.__init__(self, cur)
 
         self.conn = conn
         self.single_action = single_action
         self.host_id = 0
         self.max_host_id = self.get_max_host()
+
+        # according to HTTP spec, Retry-After can also have absolute
+        # time value, but that had not been seen yet
+        self.relative_rx = re.compile("^([0-9]{1,3})$")
+        self.holds = {} # host_id -> int time in secs
+        self.last_expiration = int(time.time() + 0.5)
         self.counter = 0
         self.healthcheck_interval = int(get_option("healthcheck_interval", "100"))
         self.healthcheck_tail = int(get_option("healthcheck_tail", "100"))
@@ -20,37 +28,77 @@ class DownloadBase(CursorWrapper):
             self.healthcheck_interval = 0
 
     def get_max_host(self):
-        self.cur.execute("""select max(id)
-from tops""")
-        row = self.cur.fetchone()
-        return row[0]
+        mx = 0
+        for _, host_id in self.host_white.items():
+            if host_id > mx:
+                mx = host_id
+
+        return mx
+
+    def get_available_hosts(self):
+        self.cond_expire()
+        avail = set(self.host_white.values())
+        return avail.difference(self.holds.keys())
+
+    def has_holds(self):
+        return len(self.holds) > 0
+
+    def add_hold(self, hostname, retry_after):
+        host_id = self.get_host_id(hostname)
+        if host_id:
+            m = self.relative_rx.match(retry_after)
+            if m:
+                relative = int(m.group(1))
+                now = self.cond_expire()
+                future = now + relative
+                old = self.holds.get(host_id)
+                if (old is None) or (old < future):
+                    self.holds[host_id] = future
+                    return True
+            else:
+                print("do not understand Retry-After: " + retry_after, file=sys.stderr)
+        else:
+            print("no hold on blacklisted " + hostname, file=sys.stderr)
+
+        return False
 
     def pop_work_item(self):
+        avail = self.get_available_hosts()
+        if not len(avail):
+            return None
+
+        last_tops = None
         while True:
-            # https://blog.2ndquadrant.com/what-is-select-skip-locked-for-in-postgresql-9-5/
-            self.cur.execute("""delete from download_queue
+            tops = [ hid for hid in avail if hid > self.host_id ]
+            tops.sort()
+            row = None
+            if len(tops) and (tops != last_tops):
+                last_tops = tops
+                # https://blog.2ndquadrant.com/what-is-select-skip-locked-for-in-postgresql-9-5/
+                self.cur.execute("""delete from download_queue
 where url_id = (
         select url_id
         from download_queue
-        where host_id > %s
+        where host_id = any(%s)
         order by host_id, priority, url_id
         for update skip locked
         limit 1
 )
-returning url_id""", (self.host_id,))
-            row = self.cur.fetchone()
-            if row:
-                self.host_id += 1
-                if self.host_id >= self.max_host_id:
-                    self.host_id = 0
+returning url_id""", (tops,))
+                row = self.cur.fetchone()
+                if row:
+                    self.host_id += 1
+                    if self.host_id >= self.max_host_id:
+                        self.host_id = 0
 
-                if self.healthcheck_interval > 0:
-                    self.counter += 1
-                    if (self.counter >= self.healthcheck_tail) and not(self.counter % self.healthcheck_interval):
-                        self.healthcheck()
+                    if self.healthcheck_interval > 0:
+                        self.counter += 1
+                        if (self.counter >= self.healthcheck_tail) and not(self.counter % self.healthcheck_interval):
+                            self.healthcheck()
 
-                return row
-            else:
+                    return row
+
+            if row is None:
                 if self.host_id > 0:
                     self.host_id = 0
                 else:
@@ -72,13 +120,50 @@ from parse_queue""")
         self.cur.execute("""notify parse_ready""")
 
     def wait(self):
+        timeout = self.get_interval()
         self.cur.execute("""listen download_ready""")
-        print("waiting for notification...")
-        select.select([self.conn], [], [])
+        if timeout is None:
+            print("waiting for notification...")
+            select.select([self.conn], [], [])
+        else:
+            print("waiting for %d second(s)..." % timeout)
+            select.select([self.conn], [], [], timeout)
+
         self.conn.poll()
         print("got %d notification(s)" % (len(self.conn.notifies),))
         while self.conn.notifies:
             self.conn.notifies.pop()
+
+    def get_interval(self):
+        # do not update holds from the time get_available_hosts had
+        # been called - if there was a hold then, we want to see it
+        # and get a finite interval
+        mn = None
+        for _, exp in self.holds.items():
+            if (mn is None) or (mn > exp):
+                mn = exp
+
+        if mn is None:
+            return None
+
+        now = int(time.time() + 0.5)
+        interval = mn - now
+        return interval if interval > 0 else 1
+
+    def cond_expire(self):
+        now = int(time.time() + 0.5)
+        if now > self.last_expiration:
+            expired_hosts = []
+            for host_id, exp in self.holds.items():
+                if exp < now:
+                    expired_hosts.append(host_id)
+
+            for host_id in expired_hosts:
+                del self.holds[host_id]
+
+            self.last_expiration = now
+
+        return now
 
     def healthcheck(self):
         print("checking success rate...", file=sys.stderr)
